Corrected the addresses for pages within DetermineActivity.java Added some basic color Fixed PNGs for workout and foot to be transparent In ppt added screenshots of all pages

+++ app/src/main/java/com/example/fitnessapplication402/CalcDistance.java
@@ -1,7 +1,7 @@
 package com.example.fitnessapplication402;
 
 public class CalcDistance {
-    private int dCalced;
+    private double dCalced;
 
     public CalcDistance(){
         dCalced = 0;
@@ -9,7 +9,7 @@ public class CalcDistance {
     public void CalcTotalDistance(int steps){
         dCalced = ((29 * steps) / 12) / 5280;
     }
-    public Integer getTotalDistance(){
+    public Double getTotalDistance(){
         return dCalced;
     }
 }
